requirements.txt
Django==1.7.4
pylint==1.4.1
pytz==2014.10
Pillow==2.7.0
astroid==1.3.4
colorama==0.3.3
django-appconf==0.6
django-autoslug==1.7.2
django-model-utils==2.2
django-podcasting==0.10
django-taggit==0.12.2
logilab-common==0.63.2
six==1.9.0
